1125 rename timeseries file
1125 update timeseries file ref in main.js
