+++ src/js/dp/search/_timeseries_handler.js
